extracted vTableInfo to its own file
